+++ libs/domain/checkout/src/services/data/checkout-data.service.ts
@@ -1,5 +1,5 @@
 import { Observable } from 'rxjs';
-import { Checkout, CheckoutData } from '../models';
+import { CheckoutData } from '../../models';
 
 export interface CheckoutDataService {
   /**
