+++ libs/domain/checkout/src/services/data/index.ts
@@ -1,0 +1,2 @@
+export * from './checkout-data.service';
+export * from './default-checkout-data.service';
