+++ libs/domain/checkout/src/services/index.ts
@@ -1,7 +1,5 @@
 export * from './checkout.providers';
 export * from './checkout.service';
+export * from './data';
 export * from './default-checkout.service';
-export * from './default-checkout-data.service';
-export * from './checkout-data.service';
-export * from './default-checkout-state.service';
-export * from './checkout-state.service';
+export * from './state';
